+++ integrations/webchat/agent-assist/agent-verify/client/javascript/connectors/genesys-agent-assist-verify.html
@@ -1,0 +1,73 @@
+<!-- 
+//
+// Copyright (c) IBM Corporation and contributors. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+-->
+
+<!DOCType html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <title>Agent Assistance Genesys</title>
+
+  <link rel="stylesheet" href="../agent-verify.css" />
+  <script type="module" src="https://1.www.s81c.com/common/carbon/web-components/version/v1.21.0/textarea.min.js"></script>
+  <script type="module" src="https://1.www.s81c.com/common/carbon/web-components/version/v1.21.0/button.min.js"></script>
+  <script src="https://sdk-cdn.mypurecloud.com/client-apps/1.3.0/purecloud-client-app-sdk.js"></script>
+  <script src="https://sdk-cdn.mypurecloud.com/javascript/123.0.0/purecloud-platform-client-v2.min.js"></script>
+  <script src="https://unpkg.com/string-similarity/umd/string-similarity.min.js"></script>
+
+  <style>
+    html, body {
+      padding: 0;
+      margin: 0;
+      width: 100%;
+      height: 100%;
+      overflow: hidden;
+    }
+
+    .DocumentContainer {
+      display: flex;
+      height: 100%;
+    }
+
+    .Panel {
+      width: 600px;
+      height: 100%;
+    }
+
+    #AgentAssistContainer {
+      padding: 16px;
+      width: 600px;
+      height: 100%;
+    }
+  </style>
+
+</head>
+
+<body>
+  <div class="DocumentContainer">
+    <div class="Panel">
+      <div id="AgentAssistContainer"></div>
+    </div>
+  </div>
+  
+  <!-- This will load the framework for enabling agent verification in web chat. -->
+  <script src="../agent-verify.js"></script>
+
+</body>
+
+</html>
