+++ integrations/webchat/agent-assist/agent-verify/client/javascript/connectors/simulator-agent-assist-verify.html
@@ -8,6 +8,7 @@
   <link rel="stylesheet" href="../agent-verify.css" />
   <script type="module" src="https://1.www.s81c.com/common/carbon/web-components/version/v1.21.0/textarea.min.js"></script>
   <script type="module" src="https://1.www.s81c.com/common/carbon/web-components/version/v1.21.0/button.min.js"></script>
+  <script src="https://unpkg.com/string-similarity/umd/string-similarity.min.js"></script>
 
   <style>
     html, body {
@@ -102,7 +103,6 @@
 
 <!-- This will load the framework for enabling agent verification in web chat. -->
 <script src="../agent-verify.js"></script>
-<script src="../string-similarity.min.js"></script> <!-- TODO: Can we use the npm version? -->
 
 </body>
 </html>
